requirements.txt
dj-database-url==0.3.0
Django==1.8.5
django-bootstrap3==6.2.2
django-haystack==2.4.1
djangorestframework==3.3.2
googlemaps==2.3
numpy==1.10.1
Pillow==3.0.0
requests==2.6.0
wheel==0.24.0
whitenoise==2.0.4
